+++ redis-server/src/main/java/site/hnfy258/command/impl/zset/Zcard.java
@@ -45,7 +45,7 @@ public class Zcard implements Command {
             throw new IllegalArgumentException("Key is not a sorted set");
         }
 
-        return RespInteger.valueOf(((RedisZset) data).size());
+        return RespInteger.valueOf((int) ((RedisZset) data).size());
     }
 
     @Override
